requirements.txt
Flask
mistralai[agents]
line-bot-sdk
black
